requirements.txt
python-chess==0.27.3
requests==2.21.0
tensorflow==1.10.0
keras==2.2.4
pandas==0.23.4
numpy==1.15.1